User signup, signin, login and logout operations. jwt and cookies. Authorization middleware.

+++ server/controllers/memories.js
@@ -1,10 +1,11 @@
 import mongoose from 'mongoose'
 import Memory from '../models/memory.js'
 
+
 export const getAll = async (req, res, next) => {
     try {
-        const datas = await Memory.find()
-        res.status(200).json(datas)
+        const data = await Memory.find()
+        res.status(200).json(data)
     } catch (err) {
         res.status(404).json({ message: err.message })
     }
@@ -25,19 +26,27 @@ export const getOne = async (req, res, next) => {
 
 export const createOne = async (req, res, next) => {
     try {
+        const createrId = req.createrId
+
         const formData = req.body
-        const newData = await Memory.create(formData)
+        const newData = await Memory.create({...formData, createrId})
         res.status(201).json({newData, message: 'New Data has been created.'})
     } catch (err) {
-        res.json({ message: err.message })
+        res.status(404).json({message: err.message})
     }
 }
 
 export const updateOne = async (req, res, next) => {
     try {
         const id = req.params.id
+        const createrId = req.createrId
+
         if (!mongoose.Types.ObjectId.isValid(id))
-            res.status(404).json({message: 'Memory Id Is Not Valid!'})
+            return res.status(404).json({message: 'Memory Id Is Not Valid!'})
+
+        const memory = await Memory.findById(id)
+        if (memory.createrId !== createrId)
+            return res.status(403).json({message: 'Update Forbidden!'})
 
         const {title, content, image, creator} = req.body
         const updateData = await Memory.findByIdAndUpdate(id, {_id: id, title, content, image, creator}, {new: true})
@@ -50,8 +59,14 @@ export const updateOne = async (req, res, next) => {
 export const deleteOne = async (req, res, next) => {
     try {
         const id = req.params.id
+        const createrId = req.createrId
+
         if (!mongoose.Types.ObjectId.isValid(id))
-            res.status(200).json({message: 'Memory Id Is Not Valid!'})
+            return res.status(404).json({message: 'Memory Id Is Not Valid!'})
+
+        const memory = await Memory.findById(id)
+        if (memory.createrId !== createrId)
+            return res.status(403).json({message: 'Delete Forbidden!'})
 
         await Memory.findByIdAndDelete(id)
         res.status(200).json({message: 'Memory Has Been Deleted'})
